requirements.txt
mysql-connector-python==8.0.24
sqlparse==0.4.1
astropy==4.2.1
pandas==1.2.4
